backend/requirements.txt
# Python dependencies for token refresh scripts
requests>=2.28.0
python-dotenv>=1.0.0
